registers

+++ compile.py
@@ -83,8 +83,9 @@ def main():
         f.write(asm.encode())
 
     os.system(assemble(config.__fileoutput__))
-    os.system(link(config.__fileoutput__,config.__fileoutput__))
-    os.remove(config.__fileoutput__+".o")
+    if(config.__executable__):
+        os.system(link(config.__fileoutput__,config.__fileoutput__))
+        os.remove(config.__fileoutput__+".o")
 
     if(not config.__tonasm__):
         os.remove(config.__fileoutput__+".asm")
